Refactor de la aplicacion y la clase main.
