--- src/main/java/com/example/demo/PruebaMasterCode.java
@@ -1,4 +1,0 @@
-package com.example.demo;
-
-public class PruebaMasterCode {
-}
