+++ src/main/java/com/example/mastercode/Master_Code_App.java
@@ -1,13 +1,13 @@
-package com.example.demo;
+package com.example.mastercode;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class Master_Code_App {
 
 	public static void main(String[] args) {
-		SpringApplication.run(DemoApplication.class, args);
+		SpringApplication.run(Master_Code_App.class, args);
 		System.out.print("Hola mundo");
 		System.out.print("Se inserta esta linea de codigo");
 		System.out.print("Wilmar inserta ewsta line de código");
